feat: implement display and comparison
I added the code to a known working emulator in the repository, and use
it in my code to compare the state of our emulators after each step.
This allowed me to verify that my emulator is working up until the first
infinite loop waiting for an interrupt.

I also modified the part of the code responsible for displaying the code
with raylib. I overlooked the fact that pixel data in video RAM take
only one bit. Now the implementation is correct and you can see the
correct screen!